fix(api): patch analytics
Ref. metriport/metriport#799
