+++ packages/api/src/routes/ehr/athenahealth/index.ts
@@ -1,5 +1,4 @@
 import Router from "express-promise-router";
-import { patientAuthorization } from "../../middlewares/patient-authorization";
 import { processPatientRoute, processDocumentRoute } from "./auth/middleware";
 import patient from "./patient";
 import chart from "./chart";
@@ -11,12 +10,7 @@ const routes = Router();
 
 routes.use("/patient", patient);
 routes.use("/chart", chart);
-routes.use(
-  "/medical/v1/patient/:id",
-  processPatientRoute,
-  patientAuthorization("query"),
-  medicalPatient
-);
+routes.use("/medical/v1/patient", processPatientRoute, medicalPatient);
 routes.use("/medical/v1/document", processDocumentRoute, medicalDocument);
 routes.use("/settings", settings);
