+++ packages/api/src/routes/medical/facility.ts
@@ -1,18 +1,86 @@
 import NotFoundError from "@metriport/core/util/error/not-found";
+import { OrgType } from "@metriport/core/domain/organization";
 import { Request, Response } from "express";
 import Router from "express-promise-router";
 import status from "http-status";
 import { deleteFacility } from "../../command/medical/facility/delete-facility";
 import { getFacilities } from "../../command/medical/facility/get-facility";
 import { updateFacility } from "../../command/medical/facility/update-facility";
+import { createFacility } from "../../command/medical/facility/create-facility";
 import { getETag } from "../../shared/http";
 import { requestLogger } from "../helpers/request-logger";
 import { getFacilityInfoOrFail } from "../middlewares/facility-authorization";
-import { asyncHandler } from "../util";
+import { asyncHandler, getCxIdOrFail } from "../util";
 import { dtoFromModel } from "./dtos/facilityDTO";
 import { facilityUpdateSchema } from "./schemas/facility";
+import { facilityAuthorization } from "../middlewares/facility-authorization";
+import { Config } from "../../shared/config";
+import { facilityCreateSchema } from "./schemas/facility";
+import { getOrganization } from "../../command/medical/organization/get-organization";
+import { createOrganization } from "../../command/medical/organization/create-organization";
 
 const router = Router();
+
+/** ---------------------------------------------------------------------------
+ * POST /facility
+ *
+ * Creates a new facility.
+ *
+ * @return {FacilityDTO} The facility.
+ */
+router.post(
+  "/",
+  requestLogger,
+  asyncHandler(async (req: Request, res: Response) => {
+    const cxId = getCxIdOrFail(req);
+    const facilityData = facilityCreateSchema.parse(req.body);
+
+    if (Config.isSandbox()) {
+      const existingOrg = await getOrganization({ cxId });
+      if (!existingOrg) {
+        await createOrganization({
+          cxId,
+          data: {
+            name: facilityData.name,
+            type: OrgType.ambulatory,
+            location: facilityData.address,
+          },
+        });
+      }
+    }
+
+    const facility = await createFacility({
+      cxId,
+      data: {
+        ...facilityData,
+        tin: facilityData.tin ?? undefined,
+        active: facilityData.active ?? undefined,
+      },
+    });
+
+    return res.status(status.CREATED).json(dtoFromModel(facility));
+  })
+);
+
+/** ---------------------------------------------------------------------------
+ * GET /facility
+ *
+ * Gets all of the facilities associated with this account.
+ *
+ * @return {FacilityDTO} The list of facilities.
+ */
+router.get(
+  "/",
+  requestLogger,
+  asyncHandler(async (req: Request, res: Response) => {
+    const cxId = getCxIdOrFail(req);
+
+    const facilities = await getFacilities({ cxId });
+
+    const facilitiesData = facilities.map(dtoFromModel);
+    return res.status(status.OK).json({ facilities: facilitiesData });
+  })
+);
 
 /** ---------------------------------------------------------------------------
  * PUT /facility/:id
@@ -22,7 +90,8 @@ const router = Router();
  * @return {FacilityDTO} The updated facility.
  */
 router.put(
-  "/",
+  "/:id",
+  facilityAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, id: facilityId } = getFacilityInfoOrFail(req);
@@ -51,7 +120,8 @@ router.put(
  * @return {FacilityDTO} The facility.
  */
 router.get(
-  "/",
+  "/:id",
+  facilityAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, id: facilityId } = getFacilityInfoOrFail(req);
@@ -71,7 +141,8 @@ router.get(
  * @return 204 if successful.
  */
 router.delete(
-  "/",
+  "/:id",
+  facilityAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, id: facilityId } = getFacilityInfoOrFail(req);
