+++ packages/api/src/routes/medical/patient.ts
@@ -1,4 +1,8 @@
-import { consolidationConversionType } from "@metriport/api-sdk";
+import {
+  consolidationConversionType,
+  patientCreateSchema,
+  demographicsSchema,
+} from "@metriport/api-sdk";
 import { GetConsolidatedQueryProgressResponse } from "@metriport/api-sdk/medical/models/patient";
 import { mrFormat } from "@metriport/core/domain/conversion/fhir-to-medical-record";
 import { MAXIMUM_UPLOAD_FILE_SIZE } from "@metriport/core/external/aws/lambda-logic/document-uploader";
@@ -38,10 +42,126 @@ import { getPatientInfoOrFail } from "../middlewares/patient-authorization";
 import { asyncHandler, getFrom, getFromQueryAsBoolean } from "../util";
 import { dtoFromModel } from "./dtos/patientDTO";
 import { bundleSchema, getResourcesQueryParam } from "./schemas/fhir";
-import { patientUpdateSchema, schemaUpdateToPatientData } from "./schemas/patient";
+import {
+  patientUpdateSchema,
+  schemaUpdateToPatientData,
+  schemaDemographicsToPatientData,
+  schemaCreateToPatientData,
+} from "./schemas/patient";
 import { cxRequestMetadataSchema } from "./schemas/request-metadata";
+import { patientAuthorization } from "../middlewares/patient-authorization";
+import {
+  getPatientOrFail,
+  matchPatient,
+  getPatients,
+} from "../../command/medical/patient/get-patient";
+import { createPatient, PatientCreateCmd } from "../../command/medical/patient/create-patient";
+import { getSandboxPatientLimitForCx } from "../../domain/medical/get-patient-limit";
+import { PatientModel as Patient } from "../../models/medical/patient";
+import { getFromQueryOrFail, getCxIdOrFail } from "../util";
 
 const router = Router();
+
+/** ---------------------------------------------------------------------------
+ * POST /patient
+ *
+ * Creates the patient corresponding to the specified facility at the
+ * customer's organization if it doesn't exist already.
+ *
+ * @param  req.query.facilityId The ID of the Facility the Patient should be associated with.
+ * @return The newly created patient.
+ */
+router.post(
+  "/",
+  requestLogger,
+  asyncHandler(async (req: Request, res: Response) => {
+    const cxId = getCxIdOrFail(req);
+    const facilityId = getFromQueryOrFail("facilityId", req);
+    const rerunPdOnNewDemographics = stringToBoolean(
+      getFrom("query").optional("rerunPdOnNewDemographics", req)
+    );
+    const forceCommonwell = stringToBoolean(getFrom("query").optional("commonwell", req));
+    const forceCarequality = stringToBoolean(getFrom("query").optional("carequality", req));
+    const payload = patientCreateSchema.parse(req.body);
+
+    if (Config.isSandbox()) {
+      // limit the amount of patients that can be created in sandbox mode
+      const numPatients = await Patient.count({ where: { cxId } });
+      const patientLimit = await getSandboxPatientLimitForCx(cxId);
+      if (numPatients >= patientLimit) {
+        return res.status(status.BAD_REQUEST).json({
+          message: `Cannot create more than ${Config.SANDBOX_PATIENT_LIMIT} patients in Sandbox mode!`,
+        });
+      }
+    }
+
+    const patientCreate: PatientCreateCmd = {
+      ...schemaCreateToPatientData(payload),
+      cxId,
+      facilityId,
+    };
+
+    const patient = await createPatient({
+      patient: patientCreate,
+      rerunPdOnNewDemographics,
+      forceCommonwell,
+      forceCarequality,
+    });
+
+    return res.status(status.CREATED).json(dtoFromModel(patient));
+  })
+);
+
+/** ---------------------------------------------------------------------------
+ * GET /patient
+ *
+ * Gets all patients corresponding to the specified facility at the customer's organization.
+ *
+ * @param   req.cxId              The customer ID.
+ * @param   req.query.facilityId  The ID of the facility the user patient is associated with (optional).
+ * @return  The customer's patients associated with the given facility.
+ */
+router.get(
+  "/",
+  requestLogger,
+  asyncHandler(async (req: Request, res: Response) => {
+    const cxId = getCxIdOrFail(req);
+    const facilityId = getFrom("query").optional("facilityId", req);
+
+    const patients = await getPatients({ cxId, facilityId: facilityId });
+
+    const patientsData = patients.map(dtoFromModel);
+    return res.status(status.OK).json({ patients: patientsData });
+  })
+);
+
+/** ---------------------------------------------------------------------------
+ * POST /patient/match
+ *
+ * Searches for a patient previously created at Metriport, based on a demographic data. Returns the matched patient, if it exists.
+ *
+ * @return The matched patient.
+ * @throws NotFoundError if the patient does not exist.
+ */
+router.post(
+  "/match",
+  requestLogger,
+  asyncHandler(async (req: Request, res: Response) => {
+    const cxId = getCxIdOrFail(req);
+    const payload = demographicsSchema.parse(req.body);
+
+    const patientData = schemaDemographicsToPatientData(payload);
+
+    const patient = await matchPatient({ cxId, ...patientData });
+
+    if (patient) {
+      // Authorization
+      await getPatientOrFail({ cxId, id: patient.id });
+      return res.status(status.OK).json(dtoFromModel(patient));
+    }
+    throw new NotFoundError("Cannot find patient");
+  })
+);
 
 /** ---------------------------------------------------------------------------
  * PUT /patient/:id
@@ -53,7 +173,8 @@ const router = Router();
  * @return The patient to be updated
  */
 router.put(
-  "/",
+  "/:id",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, id, patient } = getPatientInfoOrFail(req);
@@ -101,7 +222,8 @@ router.put(
  * @return  The customer's patients associated with the given facility.
  */
 router.get(
-  "/",
+  "/:id",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { patient } = getPatientInfoOrFail(req);
@@ -121,7 +243,8 @@ router.get(
  * @return 204 No Content
  */
 router.delete(
-  "/",
+  "/:id",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, id } = getPatientInfoOrFail(req);
@@ -155,7 +278,8 @@ router.delete(
  * @return Patient's consolidated data.
  */
 router.get(
-  "/consolidated",
+  "/:id/consolidated",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { patient } = getPatientInfoOrFail(req);
@@ -189,7 +313,8 @@ router.get(
  * @returns all consolidated queries for the patient that have been triggered.
  */
 router.get(
-  "/consolidated/query",
+  "/:id/consolidated/query",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { patient } = getPatientInfoOrFail(req);
@@ -233,7 +358,8 @@ const medicalRecordFormatSchema = z.enum(mrFormat);
  * @return status for querying the Patient's consolidated data.
  */
 router.post(
-  "/consolidated/query",
+  "/:id/consolidated/query",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, id: patientId } = getPatientInfoOrFail(req);
@@ -279,7 +405,8 @@ router.post(
  * @throws NotFoundError if the medical record summary does not exist.
  */
 router.get(
-  "/medical-record",
+  "/:id/medical-record",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, id: patientId } = getPatientInfoOrFail(req);
@@ -304,7 +431,8 @@ router.get(
  * @return JSON containing the status of the patient's medical record summary.
  */
 router.get(
-  "/medical-record-status",
+  "/:id/medical-record-status",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, id: patientId } = getPatientInfoOrFail(req);
@@ -317,7 +445,12 @@ router.get(
  * POST /patient/:id/consolidated
  * @deprecated use the PUT version of this endpoint
  */
-router.post("/consolidated", requestLogger, asyncHandler(putConsolidated));
+router.post(
+  "/:id/consolidated",
+  patientAuthorization("params"),
+  requestLogger,
+  asyncHandler(putConsolidated)
+);
 /** ---------------------------------------------------------------------------
  * PUT /patient/:id/consolidated
  *
@@ -328,7 +461,12 @@ router.post("/consolidated", requestLogger, asyncHandler(putConsolidated));
  * @param req.body The FHIR Bundle to create or update resources.
  * @return FHIR Bundle with operation outcome.
  */
-router.put("/consolidated", requestLogger, asyncHandler(putConsolidated));
+router.put(
+  "/:id/consolidated",
+  patientAuthorization("params"),
+  requestLogger,
+  asyncHandler(putConsolidated)
+);
 async function putConsolidated(req: Request, res: Response) {
   // Limit the payload size that can be created
   const contentLength = req.headers["content-length"];
@@ -358,7 +496,8 @@ async function putConsolidated(req: Request, res: Response) {
  * @param req.query.dateTo Optional end date that resources will be filtered by (inclusive).
  */
 router.get(
-  "/consolidated/count",
+  "/:id/consolidated/count",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { patient } = getPatientInfoOrFail(req);
@@ -393,7 +532,8 @@ router.get(
  * @return The patient's facility matches.
  */
 router.get(
-  "/facility-matches",
+  "/:id/facility-matches",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { id: patientId } = getPatientInfoOrFail(req);
@@ -414,7 +554,8 @@ router.get(
  * @param req.query.requestId The ID of the request.
  */
 router.get(
-  "/consolidated/webhook",
+  "/:id/consolidated/webhook",
+  patientAuthorization("params"),
   requestLogger,
   asyncHandler(async (req: Request, res: Response) => {
     const { cxId, patient } = getPatientInfoOrFail(req);
